+++ src/clj_kondo/impl/config.clj
@@ -43,6 +43,7 @@
               :redundant-declare {:level :warning}
               :var-same-name-except-case {:level :warning}
               :constant-condition {:level :warning}
+              :unreachable-code {:level :warning}
               :datalog-syntax {:level :error}
               :unbound-destructuring-default {:level :warning}
               :used-underscored-binding {:level :off}